Move types into react-query/shared directory, move new procedure infer types intos inferReactQueryProcedure
