+++ packages/react-query/src/shared/index.ts
@@ -10,6 +10,7 @@ export * from './hooks/createHooksInternal';
 export * from './queryClient';
 export * from './types';
 export * from './hooks/types';
+export * from './polymorphism';
 
 export {
   /**
